fixed gameId list
